+++ lib/MIP/Recipes/Analysis/Frequency_filter.pm
@@ -220,7 +220,7 @@ sub analysis_frequency_filter {
         {
             max_cores_per_node   => $active_parameter_href->{max_cores_per_node},
             modifier_core_number => scalar @{ $file_info_href->{contigs} },
-            module_core_number   => $core_number,
+            recipe_core_number   => $core_number,
         }
     );