+++ lib/MIP/Recipes/Analysis/Tiddit.pm
@@ -199,7 +199,7 @@ sub analysis_tiddit {
         {
             max_cores_per_node   => $max_cores_per_node,
             modifier_core_number => $modifier_core_number,
-            module_core_number   => $core_number,
+            recipe_core_number   => $core_number,
         }
     );